fix useFavoriteArticle
